UpdateChecker: Create a new Helper: Architecture helper To get the architecture of the player.
